+++ Prefs/TodayViewController.swift
@@ -30,21 +30,15 @@ class TodayViewController: UIViewController, NCWidgetProviding {
         
     }()
     
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        let layout = collectionView.collectionViewLayout as! UICollectionViewFlowLayout
-        layout.itemSize = CGSize(width: (UIScreen.main.bounds.size.width - 60)/3, height: view.bounds.size.height/11*4)
-        layout.minimumLineSpacing = view.bounds.size.height/11
-        extensionContext?.widgetLargestAvailableDisplayMode = .expanded
-    }
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view from its nib.
         preferredContentSize = CGSize(width: UIScreen.main.bounds.size.width-16, height: 110)
         view.addSubview(collectionView)
-        
         let container = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: "group.chengluffy.app-Prefs")
         let realmURL = container!.appendingPathComponent("defualt.realm")
+        
+        extensionContext?.widgetLargestAvailableDisplayMode = .expanded
         
         Realm.Configuration.defaultConfiguration.fileURL = realmURL
         realm = try! Realm()
@@ -59,18 +53,25 @@ class TodayViewController: UIViewController, NCWidgetProviding {
         if activeDisplayMode == .compact {
             preferredContentSize = maxSize
             collectionView.frame.size = maxSize
+            let layout = collectionView.collectionViewLayout as! UICollectionViewFlowLayout
+            layout.itemSize = CGSize(width: (UIScreen.main.bounds.size.width - 60)/3, height:maxSize.height/11*4)
+            print(maxSize.height/11*4)
+            layout.minimumLineSpacing = maxSize.height/11
         } else {
             var height: CGFloat?
             if (realm?.objects(Setting.self).filter("isDeleted = false").count)! > 3 {
-                let temp = view.bounds.size.height/11
-                let lines = CGFloat(realm!.objects(Setting.self).filter("isDeleted = false").count - 1)/3
-                height = CGFloat(temp * 6 + (temp * 5 * lines))
+                let layout = collectionView.collectionViewLayout as! UICollectionViewFlowLayout
+                let temp = layout.itemSize.height / 4
+                print(temp)
+                let lines: Int = (realm!.objects(Setting.self).filter("isDeleted = false").count - 1)/3
+                height = CGFloat(temp*6 + (temp * 5 * CGFloat(lines)))
             } else {
                 height = 110
             }
-            preferredContentSize = CGSize(width: maxSize.width, height: height!)
+            height = maxSize.height > height! ? height! : maxSize.height
+            print(height ?? "nil")
             collectionView.frame.size = CGSize(width: maxSize.width, height: height!)
-            
+            preferredContentSize = CGSize(width: maxSize.width, height: height!)
         }
         
         print(maxSize)